--- client/src/components/Form/Input.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-export function Input(props) {
-  return (
-    <div className="form-group">
-      <input className="form-control" {...props} />
-    </div>
-  );
-}
