--- client/src/components/Grid/Container.js
@@ -1,6 +1,0 @@
-import React from "react";
-
-// This Container component allows us to use a bootstrap container without worrying about class names
-export function Container({ fluid, children }) {
-  return <div className={`container${fluid ? "-fluid" : ""}`}>{children}</div>;
-}
