--- client/src/components/Grid/Row.js
@@ -1,6 +1,0 @@
-import React from "react";
-
-// This Row component lets us use a bootstrap row without having to think about class names
-export function Row({ fluid, children }) {
-  return <div className={`row${fluid ? "-fluid" : ""}`}>{children}</div>;
-}
