--- client/src/components/Jumbotron/Jumbotron.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-function Jumbotron({ children }) {
-  return (
-    <div
-      style={{ height: 300, clear: "both", paddingTop: 120, textAlign: "center" }}
-      className="jumbotron"
-    >
-      {children}
-    </div>
-  );
-}
-
-export default Jumbotron;
