+++ client/src/components/Jumbotron/index.js
@@ -1,1 +1,14 @@
-export { default } from "./Jumbotron";
+import React from "react";
+
+function Jumbotron({ children }) {
+  return (
+    <div
+      style={{ height: 300, clear: "both", paddingTop: 120, textAlign: "center" }}
+      className="jumbotron"
+    >
+      {children}
+    </div>
+  );
+}
+
+export default Jumbotron;
