--- client/src/components/List/ListItem.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export function ListItem({ children }) {
-  return <li className="list-group-item">{children}</li>;
-}
