--- client/src/components/Nav/Nav.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-function Nav() {
-  return (
-    <nav className="navbar navbar-expand-lg navbar-dark bg-primary">
-      <a className="navbar-brand" href="/">
-        React Reading List
-      </a>
-    </nav>
-  );
-}
-
-export default Nav;
