+++ client/src/components/Nav/index.js
@@ -1,1 +1,13 @@
-export { default } from "./Nav";
+import React from "react";
+
+function Nav() {
+  return (
+    <nav className="navbar navbar-expand-lg navbar-dark bg-primary">
+      <a className="navbar-brand" href="/">
+        React Reading List
+      </a>
+    </nav>
+  );
+}
+
+export default Nav;
