--- client/src/pages/Books/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Books.js";
