+++ models/book.js
@@ -3,8 +3,10 @@ const Schema = mongoose.Schema;
 
 const bookSchema = new Schema({
   title: { type: String, required: true },
-  author: { type: String, required: true },
-  synopsis: String,
+  authors: { type: String, required: true },
+  description: String,
+  image: String,
+  link: String,
   date: { type: Date, default: Date.now }
 });
